1_test_unfold/requirements.txt
asgiref==3.8.1
Django==5.2
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-js-asset==3.1.2
django-unfold==1.0.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
